fix(schema): use BigInt for Posicao.idPacote and SyncCursor.lastIdPacote
Both columns are already BIGINT in the DB (migration 0003 used
BIGINT for id_pacote and last_id_pacote), but the GraphQL schema
declared them as Int. Sascar returns idPacote values that exceed
2^31 (e.g., 15016348018), which the GraphQL Int scalar rejects
with 'Int cannot represent non 32-bit signed integer value',
breaking posicoesRecentes end-to-end.

Switch both fields to the BigInt scalar (added in commit 3aeac17 for
idEquipamento). Serialized as string in JSON, preserving precision.

+++ src/graphql/schema.ts
@@ -54,7 +54,7 @@ export const typeDefs = gql`
   }
 
   type Posicao {
-    idPacote: Int!
+    idPacote: BigInt!
     idVeiculo: Int!
     dataPosicao: DateTime!
     dataPacote: DateTime!
@@ -70,7 +70,7 @@ export const typeDefs = gql`
   type SyncCursor {
     method: String!
     idVeiculo: Int!
-    lastIdPacote: Int
+    lastIdPacote: BigInt
     lastSyncedAt: DateTime!
   }
